add support for injection / distribution tests, start on p3 tmpltbank test (#2277)
* start on tmpltbank test

* climate

* fixes

* poke

* fixes

* make sure to use ints

* fix example for python3

* another int conversion

+++ examples/psd/estimate.py
@@ -16,7 +16,7 @@ ts = pycbc.noise.noise_from_psd(tsamples, delta_t, psd, seed=127)
 # Estimate the PSD
 ## We'll choose 4 seconds PSD samples that are overlapped 50 %
 seg_len = int(4 / delta_t)
-seg_stride = seg_len / 2
+seg_stride = int(seg_len / 2)
 estimated_psd = pycbc.psd.welch(ts,
                       seg_len=seg_len,
                       seg_stride=seg_stride)
